+++ src/SubsetSumHillClimb.java
@@ -26,10 +26,10 @@ public class SubsetSumHillClimb {
      *
      * @return double - the residue
      */
-    public static long subsetResidue(List<Integer> multiset, long sum, int reps){
-        List<Integer> currSubset = SubsetUtil.genRandomSubset(multiset);
+    public static long subsetResidue(List<Long> multiset, long sum, int reps){
+        List<Long> currSubset = SubsetUtil.genRandomSubset(multiset);
         long currResidue = SubsetUtil.getResidue(currSubset, sum);
-        List<Integer> neighbor;
+        List<Long> neighbor;
         long neighborResidue;
 
         for(int i = 0; i < reps; i++){
@@ -51,7 +51,7 @@ public class SubsetSumHillClimb {
      * @param args
      */
     public static void main(String[] args){
-        List<Integer> intList = new ArrayList<>(Arrays.asList(1, 3, 9, 2, 7, 34, 8, 2, 45, 4));
+        List<Long> intList = new ArrayList<>(Arrays.asList(1L, 3L, 9L, 2L, 7L, 34L, 8L, 2L, 45L, 4L));
         long residue = subsetResidue(intList, 15, 20);
         System.out.println(residue);
     }
